Create WIN_____SERVER_REBOOTS_V5
This is version 5 of our reboot server ...  logic was added to have the variable <<SERVER_INFO>> either continue to contain the actual parameters that this script uses, OR, if it is set to a fullpath UNC filename, then that file will be read, and used as input to this script.  This was needed to be a work-around the 1024 character length restriction of JAMS VARIABLE values.  other minor changes were added, if a service does not stop, the script will simply continue on with a reboot of the server instead of throwing an error.